\name{appendOffset}
\alias{appendOffset}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
appendOffset}
\description{
This function is used in the addtolist function for analysis of the offsets for each one of rows.}
\usage{
appendOffset(curoffsets, newoffsets)
}
%- maybe also 'usage' for other objects documented here.

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }% use one of  RShowDoc("KEYWORDS")
