\name{check}
\alias{check}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Checking the variables and functions. }
\description{ This function is used to determine a variable or function is defined or not.}
\usage{
check(x)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }% use one of  RShowDoc("KEYWORDS")

