\name{countNulls}
\alias{countNulls}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{countNulls}
\description{
This function determines the number of missing parts in a special part of the dataset.}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }% use one of  RShowDoc("KEYWORDS")

