\name{exactidentical.norowname}
\alias{exactidentical.norowname}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Checking the equality of two parts of the dataset.}
\description{
Exactidentical.norowname function is one of the best ways to test whether two parts of the dataset are exactly equal. Exactidentical.norowname function returns TRUE when two parts of the dataset are equal and it returns FALSE when two parts of the dataset are not equal.
}
\usage{
exactidentical.norowname(df1,df2)
}


\keyword{ internal }% __ONLY ONE__ keyword per line
