% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gho_data.R
\name{gho_data}
\alias{gho_data}
\title{Load data from the GHO}
\usage{
gho_data(inds, query = NULL, year_range = "numeric")
}
\arguments{
\item{inds}{A character vector of indicator codes from the GHO database. Takes any of the codes in the \code{IndicatorCode} column of the data frame generated by \code{\link[=gho_indicators]{gho_indicators()}}.}

\item{query}{A character vector of strings fitting the \href{https://www.odata.org/documentation/odata-version-2-0/uri-conventions/}{Odata protocol} that must start with \code{"$filter="}.
If one string provided, query will be recycled for each value in \code{inds}, otherwise the length of \code{inds} and \code{query} must match.}

\item{year_range}{A string specifying whether the beginning and end values of the year range should be \code{"numeric"}, the default,
or \code{"date"}.}
}
\value{
A data frame.
}
\description{
\code{gho_data()} loads data on any number of indicators from the GHO into a single data frame.
}
