% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gho_data.R
\name{gho_dimension_values}
\alias{gho_dimension_values}
\title{Load values for specific GHO dimension}
\usage{
gho_dimension_values(dim, query = NULL)
}
\arguments{
\item{dim}{A string of the dimension code from the GHO. Takes one of the codes in the \code{Code} column of the data frame generated by \code{\link[=gho_dimensions]{gho_dimensions()}}.}

\item{query}{A string fitting the \href{https://www.odata.org/documentation/odata-version-2-0/uri-conventions/}{Odata protocol} that must start with \code{"$filter="}.}
}
\value{
A data frame.
}
\description{
\code{gho_dimension_values()} provides a data frame of all \href{https://www.who.int/data/gho/info/gho-odata-api#exe2}{available values for a specific dimension}.
}
