\name{hasNullRow}
\alias{hasNullRow}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ hasNullRow}
\description{
This function checks to have the missing rows for the special part of the dataset.
}
\usage{
hasNullRow(indf)
}

\keyword{ internal }% use one of  RShowDoc("KEYWORDS")

