\name{ht}
\alias{ht}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ht is a list of Hash objects.}
\description{
This is the collection of all Hash objects with in their offest.
}
\usage{
ht
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }% __ONLY ONE__ keyword per line
