\name{identical.norowname}
\alias{identical.norowname}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Checking the equality of two rows of the dataset.}
\description{

Identical.norowname function is one of the best ways to test whether two rows of the dataset are equal. Identical.norowname function returns TRUE when two parts of the dataset are equal and it returns FALSE when two parts of the dataset are not equal.Epsilon is a threshold for determining the equality for two rows.
}
\usage{
identical.norowname(df1, df2, epsilon)
}


\keyword{ internal }% use one of  RShowDoc("KEYWORDS")

