\name{out_csv}
\alias{out_csv}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Exporting a .csv file to the special path in pc.}
\description{
This function gets a dataset and convert that to .csv file and finally save that to a special path in pc.
}
\usage{
out_csv(data_frame, direction_save)
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal }% use one of  RShowDoc("KEYWORDS")

