\name{resolveNoSignal_revised}
\alias{resolveNoSignal_revised}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Reconstructing the missing section.
}
\description{
This function finds a possible section of the dataset for repacing with the missing section.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }% use one of  RShowDoc("KEYWORDS")

