\name{slidewindow}
\alias{slidewindow}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
slidewindow}
\description{
This function divides the dataset into different parts and saves them in an object.}
\usage{
slidewindow(indf, constraintCol, windowS)
}


\keyword{ internal }% __ONLY ONE__ keyword per line
