% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gho_data.R
\name{gho_data}
\alias{gho_data}
\title{Load data from the GHO}
\usage{
gho_data(inds, query = NULL)
}
\arguments{
\item{inds}{A character vector of indicator codes from the GHO database. Takes any of the codes in the \code{IndicatorCode} column of the data frame generated by [gho_indicators()].}

\item{query}{A string fitting the [Odata protocol](https://www.odata.org/documentation/odata-version-2-0/uri-conventions/) that must start with \code{"$filter="}.}
}
\value{
A data frame.
}
\description{
\code{gho_data()} loads data on any number of indicators from the GHO into a single data frame.
}
